[WIP] Setting up initial and downloader tests
The basic groundwork. Set up RSpec, and so on.
The only thing out of ordinary here is the downloader setup. We are not
allowed to download the actual XML file when running tests, but we need
to make sure we can get that XML from the internets when running our
program.

The question is: do we even need to check it in our tests? Can we do it
just through the results of other specs? Like, if I use Nokogiri, for
instance, a reasonable expectation would be for Nokogiri to work out of
the box, without me needing to test it.