[WIP] Change play parser, start play statistics
We created a PlayParser object to get the info we wanted from inside the
play.
At first, I was sticking both the responsibility of parsing the XML and
crunching the numbers inside PlayParser, but it got too convoluted to
read. If I had a time pressure to deliver that soon, I'd be probably
done with it and would've moved on to the next thing, but I'd get back
there after 2-3 weeks and I'd have no idea of what I did.

To curb that, I created a PlayStatistics class, which sole
responsibility is to crunch the numbers from the play.